Various low-level helper functions

Various low-level helper functions used in Alt() and kform()

Usage

consolidate(S)
kill_trivial_rows(S)
include_perms(S)
kform_to_ktensor(S)

Arguments

S

Object of class spray

Details

Low-level helper functions.

Function consolidate() takes a spray object, and combines any rows that are identical up to a permutation, respecting the sign of the permutation

Function kill_trivial_rows() takes a spray object and deletes any rows with a repeated entry (which have \(k\)-forms identically zero)

Function include_perms() replaces each row of a spray object with all its permutations, respecting the sign of the permutation

Function ktensor_to_kform() coerces a \(k\)-form to a \(k\)-tensor

Value

The functions documented here all return a spray object.

Author

Robin K. S. Hankin

See also

Examples

(S <- spray(matrix(c(1,1,2,2,1,3,3,1,3,5),ncol=2,byrow=TRUE),1:5))
#>          val
#>  3 5  =    5
#>  3 1  =    4
#>  1 3  =    3
#>  2 2  =    2
#>  1 1  =    1

kill_trivial_rows(S)  # (rows 1 and 3 killed, repeated entries)
#>          val
#>  1 3  =    3
#>  3 1  =    4
#>  3 5  =    5
consolidate(S)        # (merges rows 2 and 4)
#>          val
#>  1 1  =    1
#>  2 2  =    2
#>  1 3  =   -1
#>  3 5  =    5
include_perms(S)      # returns a spray object, not alternating tensor.
#>          val
#>  5 3  =   -5
#>  3 5  =    5
#>  3 1  =    1
#>  1 3  =   -1