Elementary forms in three-dimensional space

Objects dx, dy and dz are the three elementary one-forms on three-dimensional space. These objects can be generated by running script vignettes/dx.Rmd, which includes some further discussion and technical documentation and creates file dx.rda which resides in the data/ directory.

The default print method is a little opaque for these objects. To print them more intuitively, use

options(kform_symbolic_print = "dx")

which is documented at print.Rd.

Usage

data(dx)

Details

See vignettes dx and exeyez for an extended discussion; a use-case is given in vector_cross_product.

References

M. Spivak 1971. Calculus on manifolds, Addison-Wesley

Author

Robin K. S. Hankin

See also

Examples

dx
#> An alternating linear map from V^1 to R with V=R^1:
#>        val
#>  1  =    1
hodge(dx)
#> [1] 1
hodge(dx,3)
#> An alternating linear map from V^2 to R with V=R^3:
#>          val
#>  2 3  =    1

dx  # default print method, not particularly intelligible
#> An alternating linear map from V^1 to R with V=R^1:
#>        val
#>  1  =    1
options(kform_symbolic_print = 'dx')  # shows dx dy dz
dx
#> An alternating linear map from V^1 to R with V=R^1:
#>  + dx 
dx^dz
#> An alternating linear map from V^2 to R with V=R^3:
#>  + dx^dz 
hodge(dx,3)
#> An alternating linear map from V^2 to R with V=R^3:
#>  + dy^dz 

as.function(dx)(ex)
#> [1] 1

options(kform_symbolic_print = NULL)  # revert to default