Basis vectors in three-dimensional space

Objects ex, ey and ez are the three elementary one-forms on three-dimensional space, sometimes denoted \((e_x,e_y,e_z)\). These objects can be generated by running script vignettes/ex.Rmd, which includes some further discussion and technical documentation and creates file exeyez.rda which resides in the data/ directory.

Details

See vignettes dx and ex for an extended discussion; a use-case is given in vector_cross_product.

References

M. Spivak 1971. Calculus on manifolds, Addison-Wesley

Author

Robin K. S. Hankin

See also

Examples

as.function(dx)(ex)
#> [1] 1

(X <- as.kform(matrix(1:12, nrow=4), c(1,2,7,11)))
#> An alternating linear map from V^3 to R with V=R^12:
#>             val
#>  4 8 12  =   11
#>  3 7 11  =    7
#>  2 6 10  =    2
#>  1 5  9  =    1
as.function(X)(cbind(e(2,12), e(6,12), e(10,12)))
#> [1] 2