Inner product of two kforms

Given two \(k\)-forms \(\alpha\) and \(\beta\), return the inner product \(\left\langle\alpha,\beta\right\rangle\). Here our underlying vector space \(V\) is \(\mathcal{R}^n\).

The inner product is a symmetric bilinear form defined in two stages. First, we specify its behaviour on decomposable \(k\)-forms \(\alpha=\alpha_1\wedge\cdots\wedge\alpha_k\) and \(\beta=\beta_1\wedge\cdots\wedge\beta_k\) as

$$ \left\langle\alpha,\beta\right\rangle=\det\left( \left\langle\alpha_i,\beta_j\right\rangle_{1\leq i,j\leq n}\right) $$

and secondly, we extend to the whole of \(\Lambda^k(V)\) through linearity.

Usage

kinner(o1,o2,M)

Arguments

o1,o2

Objects of class kform

M

Matrix

Value

Returns a real number

Author

Robin K. S. Hankin

Note

There is a vignette available: type vignette("kinner") at the command line.

See also

Examples

a <- (2*dx)^(3*dy)
b <- (5*dx)^(7*dy)

kinner(a,b)
#> [1] 210
det(matrix(c(2*5,0,0,3*7),2,2))  # mathematically identical, slight numerical mismatch
#> [1] 210