Elementary tensors

Creates the elementary tensors or tensor products of elementary tensors

phi(n)

Arguments

n

Vector of strictly non-negative integers

Details

If \(v_1,\ldots,v_n\) is the standard basis for \(\mathbb{R}^n\) then \(\phi_i\) is defined so that \(\phi_i(v_j)=\delta_{ij}\). phi(n) returns \(\phi_n\).

If n is a vector of strictly positive integers, then phi(n) returns the tensor cross product of \(\phi\) applied to the individual elements of n [which is a lot easier and more obvious than it sounds].

Author

Robin K. S. Hankin

Note

There is a vignette, phi

Examples

phi(6)
#> A linear map from V^1 to R with V=R^6:
#>        val
#>  6  =    1
phi(6:8)
#> A linear map from V^3 to R with V=R^8:
#>            val
#>  6 7 8  =    1

v <- sample(9)
phi(v) == Reduce("%X%",sapply(v,phi))
#> [1] TRUE