Random kforms and ktensors

Random \(k\)-form objects and \(k\)-tensors, intended as quick “get you going” examples

Usage

rform(terms=9,k=3,n=7,coeffs,ensure=TRUE)
rtensor(terms=9,k=3,n=7,coeffs)

Arguments

terms

Number of distinct terms

k,n

A \(k\)-form maps \(V^k\) to \(\mathbb{R}\), where \(V=\mathbb{R}^n\)

coeffs

The coefficients of the form; if missing use seq_len(terms)

ensure

Boolean with default TRUE meaning to ensure that the dovs() of the returned value is in fact equal to n. If FALSE, sometimes the dovs() is strictly less than n because of random sampling

Details

What you see is what you get, basically.

Note that argument terms is an upper bound, as the index matrix might contain repeats which are combined.

Value

All functions documented here return an object of class kform or ktensor.

Author

Robin K. S. Hankin

Examples

rform()
#> An alternating linear map from V^3 to R with V=R^7:
#>            val
#>  1 4 7  =    6
#>  3 5 6  =   -8
#>  2 6 7  =    3
#>  2 3 5  =    2
#>  2 3 7  =   -2
#>  1 3 4  =    4
rform() ^ rform()
#> An alternating linear map from V^6 to R with V=R^7:
#>                  val
#>  1 2 3 5 6 7  =   -6
#>  1 3 4 5 6 7  =  -28
#>  1 2 3 4 5 6  =  -21
#>  2 3 4 5 6 7  =   -7
#>  1 2 3 4 6 7  =  -24
#>  1 2 3 4 5 7  =  -45
#>  1 2 4 5 6 7  =   78

rtensor() %X% rtensor()
#> A linear map from V^6 to R with V=R^7:
#>                  val
#>  3 1 1 3 5 5  =    2
#>  6 3 2 3 5 5  =    3
#>  5 7 3 3 5 5  =    8
#>  6 2 3 3 5 5  =    4
#>  6 7 5 3 5 5  =    5
#>  6 4 7 3 5 5  =    9
#>  6 3 6 1 2 1  =    8
#>  3 1 1 1 2 1  =   16
#>  6 2 3 1 2 1  =   32
#>  6 2 5 1 2 1  =   56
#>  6 3 6 5 4 1  =    2
#>  5 7 3 1 2 1  =   64
#>  5 7 3 5 4 1  =   16
#>  6 4 7 4 1 6  =   45
#>  2 4 5 1 2 1  =   48
#>  6 7 5 2 4 2  =   35
#>  2 4 5 3 5 5  =    6
#>  6 2 5 4 1 6  =   35
#>  3 1 1 2 6 1  =   12
#>  5 7 3 2 6 1  =   48
#>  5 7 3 2 2 3  =   32
#>  6 3 2 2 6 1  =   18
#>  6 3 2 4 1 6  =   15
#>  6 3 2 1 2 1  =   24
#>  2 4 5 4 1 6  =   30
#>  6 7 5 2 6 1  =   30
#>  2 4 5 2 6 1  =   36
#>  3 1 1 2 4 2  =   14
#>  6 4 7 5 1 5  =   81
#>  6 3 6 5 1 5  =    9
#>  6 4 7 1 2 1  =   72
#>  6 4 7 2 6 1  =   54
#>  3 1 1 5 4 1  =    4
#>  2 4 5 5 1 5  =   54
#>  6 2 5 2 4 2  =   49
#>  6 2 3 4 1 6  =   20
#>  6 2 5 2 6 1  =   42
#>  6 7 5 5 1 5  =   45
#>  3 1 1 5 1 5  =   18
#>  6 3 6 2 6 1  =    6
#>  6 2 3 5 4 1  =    8
#>  6 3 2 5 1 5  =   27
#>  3 1 1 2 2 3  =    8
#>  2 4 5 5 4 1  =   12
#>  2 4 5 2 4 2  =   42
#>  6 7 5 1 2 1  =   40
#>  6 2 5 1 1 6  =   21
#>  6 4 7 2 4 2  =   63
#>  6 3 2 2 4 2  =   21
#>  6 3 6 2 4 2  =    7
#>  6 3 6 2 2 3  =    4
#>  6 2 5 5 1 5  =   63
#>  6 3 2 2 2 3  =   12
#>  6 2 3 2 6 1  =   24
#>  6 7 5 4 1 6  =   25
#>  2 4 5 1 1 6  =   18
#>  5 7 3 4 1 6  =   40
#>  6 2 3 1 1 6  =   12
#>  6 3 6 3 5 5  =    1
#>  6 2 3 5 1 5  =   36
#>  6 3 2 1 1 6  =    9
#>  6 7 5 5 4 1  =   10
#>  6 3 6 4 1 6  =    5
#>  3 1 1 4 1 6  =   10
#>  6 4 7 2 2 3  =   36
#>  2 4 5 2 2 3  =   24
#>  6 2 5 2 2 3  =   28
#>  5 7 3 1 1 6  =   24
#>  5 7 3 5 1 5  =   72
#>  6 3 6 1 1 6  =    3
#>  6 2 5 3 5 5  =    7
#>  6 7 5 2 2 3  =   20
#>  6 2 3 2 4 2  =   28
#>  3 1 1 1 1 6  =    6
#>  6 2 3 2 2 3  =   16
#>  6 4 7 1 1 6  =   27
#>  5 7 3 2 4 2  =   56
#>  6 7 5 1 1 6  =   15
#>  6 4 7 5 4 1  =   18
#>  6 3 2 5 4 1  =    6
#>  6 2 5 5 4 1  =   14

rform() ^ dx
#> An alternating linear map from V^4 to R with V=R^7:
#>              val
#>  1 2 3 7  =    2
#>  1 3 4 7  =   -1
#>  1 2 5 6  =   -4
rform() ^ dx ^ dy
#> An alternating linear map from V^5 to R with V=R^7:
#>                val
#>  1 2 3 4 6  =    1
#>  1 2 3 4 7  =   -3
#>  1 2 4 6 7  =   -4
#>  1 2 3 4 5  =    8
#>  1 2 4 5 7  =   -7