Zap small values in \(k\)-forms and \(k\)-tensors

Equivalent to zapsmall()

Usage

zap(X)
# S3 method for class 'kform'
zap(X)
# S3 method for class 'ktensor'
zap(X)

Arguments

X

Tensor or \(k\)-form to be zapped

Details

Given an object of class ktensor or kform, coefficients close to zero are ‘zapped’, i.e., replaced by ‘0’, using base::zapsmall().

Note, zap() actually changes the numeric value, it is not just a print method.

Value

Returns an object of the same class

Author

Robin K. S. Hankin

Examples

S <- rform(7)
S == zap(S)
#> [1] TRUE